FILE: relocates remaining types/game contents to data/ folder
