+++ libs/spirit-islander/util/src/index.ts
@@ -1,7 +1,11 @@
 // Data
 export * from './lib/data/adversaries';
 export * from './lib/data/boards';
+export * from './lib/data/difficulty';
+export * from './lib/data/expansions';
 export * from './lib/data/maps';
+export * from './lib/data/option';
+export * from './lib/data/players';
 export * from './lib/data/scenarios';
 export * from './lib/data/spirits';
 
@@ -10,12 +14,6 @@ export * from './lib/types/combo.interface';
 export * from './lib/types/config.interface';
 export * from './lib/types/game-setup.interface';
 export * from './lib/types/route.enum';
-
-// Game types
-export * from './lib/types/game/difficulty';
-export * from './lib/types/game/expansions';
-export * from './lib/types/game/option';
-export * from './lib/types/game/players';
 
 // Functions
 export * from './lib/functions/app-migration';
