+++ libs/spirit-islander/util/src/lib/data/difficulty.ts
