+++ libs/spirit-islander/util/src/lib/data/option.ts
