+++ libs/spirit-islander/util/src/lib/types/game-setup.interface.ts
@@ -1,7 +1,7 @@
-import type { Players } from './game/players';
-import type { Difficulty } from './game/difficulty';
+import type { Players } from '../data/players';
+import type { Difficulty } from '../data/difficulty';
 import type { Spirit } from '../data/spirits';
-import type { ExpansionName } from './game/expansions';
+import type { ExpansionName } from '../data/expansions';
 import { AdversaryLevel } from '../data/adversaries';
 import type { Scenario } from '../data/scenarios';
 import type { Map } from '../data/maps';
